fix: align facade package contracts
