+++ packages/release-drafter/tsconfig.declarations.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "noResolve": true,
+    "rootDir": "src"
+  },
+  "files": ["src/index.ts", "src/types.ts"],
+  "include": []
+}
